emotion_reconition/requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
python-dotenv
